custom for Jingli

+++ src/pgsql/extensions/bin/extensions/pgl_ddl_deploy/install.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-apt-get install -y make gcc
+if [ "$EXTENSIONS_BATCH_INSTALL" = "" ]; then
+  apt-get install -y make gcc
+fi
 wget -O pgl_ddl_deploy.tar.gz "https://github.com/enova/pgl_ddl_deploy/tarball/REL1_2_STABLE"
 tar -zxvf pgl_ddl_deploy.tar.gz
 cd enova-pgl_ddl_deploy*
@@ -10,4 +12,6 @@ cd ..
 
 #cleanup
 rm -rf enova-pgl_ddl_deploy* && rm -rf pgl_ddl_deploy.tar.gz
-apt-get remove --auto-remove -y make gcc
+if [ "$EXTENSIONS_BATCH_INSTALL" = "" ]; then
+  apt-get remove --auto-remove -y make gcc
+fi
