+++ src/pgsql/extensions/bin/extensions/pgosm/install.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-apt-get install -y make build-essential
+if [ "$EXTENSIONS_BATCH_INSTALL" = "" ]; then
+  apt-get install -y make build-essential
+fi
 cd /tmp && \
     wget -qO-  https://github.com/openstreetmap/openstreetmap-website/archive/master.tar.gz | tar xz
 
@@ -8,6 +10,8 @@ cd /tmp/openstreetmap-website-master/db/functions
 make libpgosm.so 
 cp libpgosm.so /usr/lib/postgresql/$PG_MAJOR/lib
 
-apt-get remove -y make build-essential
+if [ "$EXTENSIONS_BATCH_INSTALL" = "" ]; then
+  apt-get remove -y make build-essential
+fi
 cd /
 rm -rf /tmp/openstreetmap-website-master/
